requirements.txt
Django==1.11.4
psycopg2==2.7.3
djangorestframework==3.6.4
